Created error handler class and added async error handling.
Created the utils directory to start breaking some code out of the app.js

ExpressError is the custom error handler class, and catchAsync() is a wrapper
function. I am using that to automatically wrap async requests
in a try catch so that all errors are handled at least in some way.

Then I will be customizing the error handlers themselves and the logic
to differentiate errors between types.